requirements.txt
bs4
discord.py
lxml
markdownify
python-dotenv
regex
requests